refactor h2oImputeInput. migrate from global variable to function import
