+++ src/index.js
@@ -8,7 +8,6 @@ const flowPrelude = flowPreludeFunction();
 
 import { parseInput } from './parseInput/parseInput';
 import { jobOutput } from './jobOutput/jobOutput';
-import { imputeInput } from './imputeInput/imputeInput';
 import { util } from './util/util';
 import { routines } from './routines/routines';
 import { coreUtils } from './coreUtils/coreUtils';
@@ -75,7 +74,6 @@ import { flow } from './flow/flow';
   coreUtils();
   routines();
   util();
-  imputeInput();
   jobOutput();
   parseInput();
 }).call(this);
